refactor: move static scale/chord definitions to their own file
